[X] Finished secret 3 and 4 [X] Fixed bug where you could throw items through the wall on the left map edge [X] Red Koopas detect platform edges and turn around [X] Cleaned up areas that used "*TILESIZE" and replaced them with "<< 5" for performance [X] Fixed bug that crashed world editor if the map was smaller than the screen [X] Fixed bug in world editor that didn't save worlds correct with boxes or bosses minigames [X] Fixed bug when placing maps into the tours folder, they are not found in the world editor

+++ _src/menu.cpp
@@ -61,7 +61,7 @@ void Menu::WriteGameOptions()
 
 	if(fp != NULL)
 	{
-		fwrite(g_iVersion, sizeof(short), 4, fp);
+		fwrite(g_iVersion, sizeof(int), 4, fp);
 
 #ifdef _XBOX
